Allow anyone to log in to cardboard
