+++ cardboard/settings.py
@@ -259,8 +259,7 @@ except KeyError as e:
 SOCIAL_AUTH_PIPELINE = (
     "social_core.pipeline.social_auth.social_details",
     "social_core.pipeline.social_auth.social_uid",
-    # custom auth_allowed check in lieu of default one
-    "google_api_lib.sync_tasks.auth_allowed",
+    "social_core.pipeline.social_auth.auth_allowed",
     "social_core.pipeline.social_auth.social_user",
     "social_core.pipeline.user.get_username",
     "social_core.pipeline.user.create_user",
