+++ google_api_lib/sync_tasks.py
@@ -2,58 +2,12 @@ import logging
 from typing import List
 
 from celery import shared_task
-from django.conf import settings
-from googleapiclient.errors import Error
-from social_core.exceptions import AuthForbidden
 
 from .utils import GoogleApiClientTask
 
 logger = logging.getLogger(__name__)
 
 HUMAN_DRIVE_FOLDER_NAME = "FOLDER FOR HUMANS"
-
-
-def auth_allowed(backend, details, response, *args, **kwargs):
-    """
-    Checks if a user is allowed by checking if the user's email
-    is in the list of emails added to the Google Drive folder.
-    Part of the SOCIAL_AUTH_PIPELINE (see settings.py).
-    """
-    email = details.get("email").lower()
-    # Allow all emails if Google Drive integration is not set up
-    # or folder is world readable. Otherwise, only allow emails
-    # added to the Google Drive folder.
-    if settings.GOOGLE_API_AUTHN_INFO:
-        user_emails = get_file_user_emails(settings.GOOGLE_DRIVE_HUNT_FOLDER_ID)
-        if user_emails is None:
-            logger.warn(
-                "The Google Drive folder for the hunt is world-readable, meaning anyone can login to this Cardboard instance."
-            )
-        elif email not in user_emails:
-            raise AuthForbidden(backend)
-
-
-@shared_task(base=GoogleApiClientTask, bind=True)
-def get_file_user_emails(self, file_id) -> List[str]:
-    """
-    Returns a sorted list of emails that have access to `file_id` or None if the file is
-    world readable.
-    """
-    response = (
-        self.drive_service().files().get(fileId=file_id, fields="permissions").execute()
-    )
-
-    permissions = response["permissions"]
-
-    emails = set()
-    for perm in permissions:
-        if perm["id"] == "anyoneWithLink":
-            return None
-        email = perm["emailAddress"]
-        emails.add(email)
-        emails.add(email.lower())
-
-    return sorted(list(emails))
 
 
 @shared_task(base=GoogleApiClientTask, bind=True)
